MailService/requirements.txt
kombu
pyyaml
